# docs/swagger.yaml
definitions:
...
  util.CommonResult:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server celler server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Swagger Example API
  version: "1.0"
paths:
...
  /coupon/{id}:
    get:
      consumes:
      - application/json
      description: 获取单个优惠券的详细信息
      parameters:
      - description: id
        in: path
        name: id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.CommonResult'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.CommonResult'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/util.CommonResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.CommonResult'
      summary: 获取单个优惠券的详细信息
      tags:
      - 优惠券管理
  /coupon/create:
    post:
      consumes:
      - application/json
      description: 添加优惠券
      parameters:
      - description: 金额
        in: query
        name: amount
        type: number
      - description: 优惠码
        in: query
        name: code
        type: string
      - description: 数量
        in: query
        name: count
        type: integer
      - description: 可以领取的日期
        in: query
        name: enableTime
        type: string
      - in: query
        name: endTime
        type: string
      - in: query
        name: id
        type: integer
      - description: 可领取的会员类型：0->无限时
        in: query
        name: memberLevel
        type: integer
      - description: 使用门槛；0表示无门槛
        in: query
        name: minPoint
        type: number
      - in: query
        name: name
        type: string
      - description: 备注
        in: query
        name: note
        type: string
      - description: 每人限领张数
        in: query
        name: perLimit
        type: integer
      - description: 使用平台：0->全部；1->移动；2->PC
        in: query
        name: platform
        type: integer
      - description: 发行数量
        in: query
        name: publishCount
        type: integer
      - description: 领取数量
        in: query
        name: receiveCount
        type: integer
      - in: query
        name: startTime
        type: string
      - description: 优惠券类型；0->全场赠券；1->会员赠券；2->购物赠券；3->注册赠券
        in: query
        name: type
        type: integer
      - description: 已使用数量
        in: query
        name: useCount
        type: integer
      - description: 使用类型：0->全场通用；1->指定分类；2->指定商品
        in: query
        name: useType
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.CommonResult'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.CommonResult'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/util.CommonResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.CommonResult'
      summary: 添加优惠券
      tags:
      - 优惠券管理
  /coupon/delete/{id}:
    post:
      consumes:
      - application/json
      description: 删除优惠券
      parameters:
      - description: id
        in: path
        name: id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.CommonResult'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.CommonResult'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/util.CommonResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.CommonResult'
      summary: 删除优惠券
      tags:
      - 优惠券管理
  /coupon/list:
    get:
      consumes:
      - application/json
      description: 根据优惠券名称和类型分页获取优惠券列表
      parameters:
      - description: 名称
        in: query
        name: name
        type: string
      - description: 类型
        in: query
        name: type
        type: string
      - description: 页码
        in: query
        name: pageNum
        type: string
      - description: 数量
        in: query
        name: pageSize
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.CommonResult'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.CommonResult'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/util.CommonResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.CommonResult'
      summary: 根据优惠券名称和类型分页获取优惠券列表
      tags:
      - 优惠券管理
  /coupon/update/{id}:
    post:
      consumes:
      - application/json
      description: 修改优惠券
      parameters:
      - description: id
        in: path
        name: id
        type: string
      - description: 金额
        in: query
        name: amount
        type: number
      - description: 优惠码
        in: query
        name: code
        type: string
      - description: 数量
        in: query
        name: count
        type: integer
      - description: 可以领取的日期
        in: query
        name: enableTime
        type: string
      - in: query
        name: endTime
        type: string
      - in: query
        name: id
        type: integer
      - description: 可领取的会员类型：0->无限时
        in: query
        name: memberLevel
        type: integer
      - description: 使用门槛；0表示无门槛
        in: query
        name: minPoint
        type: number
      - in: query
        name: name
        type: string
      - description: 备注
        in: query
        name: note
        type: string
      - description: 每人限领张数
        in: query
        name: perLimit
        type: integer
      - description: 使用平台：0->全部；1->移动；2->PC
        in: query
        name: platform
        type: integer
      - description: 发行数量
        in: query
        name: publishCount
        type: integer
      - description: 领取数量
        in: query
        name: receiveCount
        type: integer
      - in: query
        name: startTime
        type: string
      - description: 优惠券类型；0->全场赠券；1->会员赠券；2->购物赠券；3->注册赠券
        in: query
        name: type
        type: integer
      - description: 已使用数量
        in: query
        name: useCount
        type: integer
      - description: 使用类型：0->全场通用；1->指定分类；2->指定商品
        in: query
        name: useType
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.CommonResult'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.CommonResult'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/util.CommonResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.CommonResult'
      summary: 修改优惠券
      tags:
      - 优惠券管理
  /couponHistory/list:
    get:
      consumes:
      - application/json
      description: 根据优惠券id，使用状态，订单编号分页获取领取记录
      parameters:
      - description: couponId
        in: query
        name: couponId
        type: string
      - description: useStatus
        in: query
        name: useStatus
        type: string
      - description: orderSn
        in: query
        name: orderSn
        type: string
      - description: 页码
        in: query
        name: pageNum
        type: string
      - description: 数量
        in: query
        name: pageSize
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.CommonResult'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.CommonResult'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/util.CommonResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.CommonResult'
      summary: 根据优惠券id，使用状态，订单编号分页获取领取记录
      tags:
      - 优惠券领取记录管理
...
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"